Rename hotel feature module to search feature module
