--- challenge/android/feature/hotels/src/androidTest/java/com/github/felipehjcosta/huchallenge/feature/hotels/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package com.github.felipehjcosta.huchallenge.feature.hotels
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.github.felipehjcosta.huchallenge.feature.hotels.test", appContext.packageName)
-    }
-}
