--- challenge/android/feature/hotels/src/androidTest/kotlin/com/github/felipehjcosta/huchallenge/di/TestApplicationComponent.kt
@@ -1,21 +1,0 @@
-package com.github.felipehjcosta.huchallenge.di
-
-import com.github.felipehjcosta.huchallenge.base.di.ApplicationComponent
-import com.github.felipehjcosta.huchallenge.base.hotels.HotelsModule
-import com.github.felipehjcosta.huchallenge.feature.hotels.HotelsUITest
-import dagger.Component
-import dagger.android.AndroidInjectionModule
-import dagger.android.support.AndroidSupportInjectionModule
-import javax.inject.Singleton
-
-@Singleton
-@Component(modules = [
-    TestApplicationModule::class,
-    TestNetworkModule::class,
-    HotelsModule::class,
-    AndroidInjectionModule::class,
-    AndroidSupportInjectionModule::class
-])
-interface TestApplicationComponent : ApplicationComponent {
-    fun inject(hotelsActivityTest: HotelsUITest)
-}
