--- challenge/android/feature/hotels/src/main/kotlin/com/github/felipehjcosta/huchallenge/feature/hotels/di/HotelsModule.kt
@@ -1,6 +1,0 @@
-package com.github.felipehjcosta.huchallenge.feature.hotels.di
-
-import dagger.Module
-
-@Module
-class HotelsModule
