--- challenge/android/feature/hotels/src/main/kotlin/com/github/felipehjcosta/huchallenge/feature/hotels/di/HotelsScope.kt
@@ -1,7 +1,0 @@
-package com.github.felipehjcosta.huchallenge.feature.hotels.di
-
-import javax.inject.Scope
-
-@Scope
-@kotlin.annotation.Retention
-annotation class HotelsScope
