+++ challenge/android/feature/search/src/androidTest/kotlin/com/github/felipehjcosta/huchallenge/HUChallengeAndroidJUnitRunner.kt
