+++ challenge/android/feature/search/src/main/kotlin/com/github/felipehjcosta/huchallenge/feature/search/di/SearchScope.kt
@@ -1,0 +1,7 @@
+package com.github.felipehjcosta.huchallenge.feature.search.di
+
+import javax.inject.Scope
+
+@Scope
+@kotlin.annotation.Retention
+annotation class SearchScope
