+++ challenge/android/feature/search/src/main/res/layout/hotels_item_layout.xml
