+++ challenge/android/feature/search/src/main/res/values/strings.xml
