+++ challenge/android/feature/search/src/test/kotlin/com/github/felipehjcosta/huchallenge/feature/TestStubApplication.kt
