+++ challenge/android/instantapp/build.gradle
@@ -1,6 +1,6 @@
 apply plugin: 'com.android.instantapp'
 
 dependencies {
-    implementation project(':feature:hotels')
+    implementation project(':feature:search')
     implementation project(':feature:base')
 }
